Feature/lecture

+++ src/main/java/com/example/csws/entity/instance/CreateInstanceRequest.java
@@ -13,5 +13,5 @@ public class CreateInstanceRequest {
     private String keyPair;
     private String address;
     private String os;
-    private int serverId;
+    private Long serverId;
 }
